Update base from 4.7 to at least 4.11 (#14)
After merging this commit, I suggest you delete `.stack-work/`, so the whole thing gets re-generated next time you run `run.sh`. The `run.sh` script has a slight brittleness, in which it assumes you've only compiled against one version of cabal, and this PR updates to a new version. 

The goal here is to get the declaration `Semigroup a => Monoid (Maybe a)`, which was added in `base` version 4.11 (released in March 2018). This will be used in the implementation of extracting the last call from an auction. Grabbing the last call from an auction will be useful when displaying self-alerts for the intended calls in the solutions: rather than describing the alerts in two places (once when making bids in an auction and once when specifying the solutions to the situations), we can just put them in the auctions, and extract them later to show the solutions with the descriptions of alerts. but first, this all requires upgrading `base` to at least 4.11. and doing that required upgrading the resolver to something new enough that it knows about base 4.11, so I set it to the latest update (LTS-20.1). 

and then while I was looking at it, I replaced `newDeal` with `mempty`, since the `DealerProg` datatype is a `Monoid`.


* Update base from 4.7 to at least 4.11, to get Semigroup a => Monoid (Maybe a)

* off-by-one error when selecting random items from a list

* alphabetize imports again

* Rephrase to clean up slightly

* whitespace tweak

* Prefer mempty instead of making our own newDeal value