Refactor situations to use the State monad (#16)
The implementation of `<~` has never sat well with me, and this is an attempt at cleaning it up a bit. You can now `wrap` a `State StdGen Situation` into a `Situations`, and can no longer do it with a `StdGen -> Situation`. This greatly simplifies `<~`, which now operates on `State StdGen (a -> b)` and gets rid of the `Optionable` typeclass entirely. and with that change, `base` was just `return` so I renamed all of those and deleted it.

Alas, I was hoping to simplify the `wrapVulDlr $ base sit` in the definition of each new `Situation`, but still can't find a way to do that (beyond renaming `base` to `return`).


* Add State StdGen Situations to Situations

* Now that we're using FlexibleInstances, get rid of the shenanigans around StdGen

* Switch the <~ syntax to use State StdGen

* remove support for SitFun

* replace base with return