Immediate responses to SMP 1D openings (#18)
Also included: `displayLastCall` formats an `Action` into a string as just the last call without any alerts, for use in the explanation of why it's the right one.

I need to firm up the constraints around what the opponents might have, when they pass after partner opens. Currently, for example, they pass with 14 HCP and a 5-card suit they could have overcalled at the 1 level. but this PR is already far too long, so that can come later. and getting it right could be pretty tricky, given that the constraints on what they might be able to overcall depends on what the most recent bid was.

An open question for me: the book defines the auction 1D-3D as showing "6 or more diamonds, with about 5 to 10 HCP," and defines the auction 1D-4D as showing "7 cards in the suit and 6-9 HCP." When would you ever bid the latter, given that in those cases you could also have bid the former? I'm unsure, so I left out those auctions and have a TODO for them. :shrug: 


* Starting on response bids to 1D openings

* Add Reverse Flannery definitions

* b1D1H

* Start the topic for responses to 1D

* Situations for responding 1M to 1D

* Add a way to mention a bid without its alert, and the Reverse Flannery Situations

* Situations for responding 3C/4C over 1D

* Add situations for notrump bids after partner opens 1D

* Use displayLastCall to simplify options

* tweak

* Add bid and situation for inverted minors over 1D

* Add 3- and 4-level pre-empts

* update Main.hs to include everything

* Rename extractCall to removeAlert

* standardize topic descriptions and filenames

* Add more 1D-2m situations

* Remove commented-out imports