advanced SMP: 1C-1H denies 4 spades and swapped meanings of 1C-1N and 1C-2H (#21)
Still left to do: the waiting bids after 1C-1H (2S is a waiting bid unless you could have bid 2D, in which case that's waiting and 2S shows diamonds). That should probably be separate files, which means it should probably be a separate PR.

Not sure I like these names: they're intended to have the bids be upper case and the extra parts be lower case, but the `alt`ernative version of `2H` looks like some kind of `Halt` action. I'll either think of better names, or merge this and think of better names later.


* Start on modified 1C responses: 1H denies spades and 1S is GF or stronger

* Swap around the 1N and 2H responses to 1C

* pr feedback

* make it easier to switch back to the default topic later