Make it easier to write commentary (#26)
I'm so pleased with the insight behind this! So much less boilerplate now. When writing commentary, instead of having a function that takes an `OutputType` and formatting each piece of the output right there, you just concatenate them all with `(.+)`, even if they naively appear to be different types, and they get formatted to whatever the output is supposed to be later! The `Commentary` type has been changed from an alias for `OutputType -> String` to its own struct containing a list of those functions (hence how the concatenation works).

The major change is in Output.hs, a handful of other files in that directory changed to match, and then all the changes within the Topics directory are just simplifying the formatting. and the next format simplification after this one could be that you can put in the `Action`s being discussed instead of writing out their bids all over again. (Not sure if that next one is a good idea: it makes it too easy to have mistakes due to copying and pasting things without updating them. but it's at least possible if you want it.)

The vestigial parts of Output.hs, including `Outputtable` and `Wrap`, have also been removed as no longer necessary. 

This is a big change, but my hope is that going commit-by-commit though the pre-squashed PR on github is easy.


* Remove unnecessary complication

* Start of changing Commentary to a monoid

* monoid instance for Commentary

* create toCommentary

* remove unnecessary parens

* Replace toCommentary with (.+)

* Simplify by moving output into Showable

* Reorder for ease of reading

* Convert OneClubResponses.hs to use (.+)

* Update displayLastCall for new Commentary format

* update Mafia.hs to use (.+)

* update MinorTransfersScott.hs to use (.+)

* update OpeningBids.hs to use (.+)

* update TwoDiamondOpeners.hs to use (.+)

* update MafiaResponses.hs to use (.+)

* update OneDiamondResponses.hs to use (.+)

* update JacobyTransfers.hs to use (.+)

* update StandardOpeners.hs to use (.+)

* cleanup: remove scaffolding

* tweak: remove unused import

* tweak: use toCommentary instead of appending the empty string

* tweak: typeclass is a single word

* tweak: order things so they only depend on things defined earlier in the file

* tweak: use concatMap instead of joining with the empty string on the map