Prepare to show actual SituationInstances via the server (#35)
The next couple steps, I suspect, are to actually implement `toHtml` for all the Showable types that don't do that yet, and maybe get rid of `output`. but things still compile, I can get the "todo" to show up, and the RNG changes as you request new situations!

I also learned that, for the Jacoby Transfers situations with 5-5 in the majors, we've been asserting that the majors must have equal length and also asserting that they must _not_ have equal length, resulting in us skipping those every time they come up. I haven't noticed before because there is so much stuff printed to stdout when rendering the LaTeX. but the fix was easy enough.


* add comment

* change port numbers to one I'm not already using elsewhere

* store the rng in the server state

* start of using rng in server, needs a way to get updated rng back

* generate should return the updated rng too

* get the StdGen out of the IORef, not using it yet

* store the updated StdGen back in the IORef

* add placeholder to render to JSON

* make toCommentary a normal function, not part of the Showable class

* move output out of the Showable class, too

* fix bug: allow equal-length majors in situations with equal length majors

* demonstrate that the RNG changes as we run