Add a Meckwell topic (#42)
It feels too wordy to describe it as "Meckwell over strong notrump," but I wanted to distinguish it from the Meckwell bidding system as a whole. I'm also unsure about some of the finer details, which seem not to have consensus online (what is the minimum length for two-suited hands? If partner has clubs and a major, how do you ask what the major is?). but this is good enough for the moment.


* define soundHolding

* define Meckwell bids, not used yet

* tweaks to business logic

* remember to alert the alertable bids

* start of meckwell topic

* tweak: it's very unlikely we'll bid Meckwell as a passed hand

* coming along

* possibly done

* 2-suited hands shouldn't be 3-suited without nuance

* tweak: 6-4 is not necessarily a 2-suited shape

* actually use the small-caps for notrump

* clarify explanation to remind you what partner is showing

* remove duplicated code