More refactoring (#59)
If this were a professional project, I would have split the work on SMP 2D out from the rest. but it's just me working on this, it's still in a usable state, and I'm about to continue work on that further. Not worth the effort to organize the commits into proper PRs.

The refactoring changes include:
- renamed Auction.hs to Action.hs, since Actions turned out to be way more important than Auctions.
- defined the new combinator `forEach`
- added `forbidAll` as syntactic sugar around `flip forEach forbid`


* reorder definitions in Auction.hs

* rename Auction.hs -> Action.hs

* very start of SMP 2D bids

* start moving bids to the Bids/ directory

* tweak names in impliesThat

* define and use forEach instead of mapM_

* define and use forbidAll

* tweak