Syntactic sugar when ensuring one hand is unpassed (#62)
This was coming up often enough I should make some wrappers around it. I still suspect I've got the wrong data structures around this: there shouldn't be a need for `stdWrap` (and its NW and SE variations) in addition to `wrapVulDlr` (and its NW and SE variations). but that's a problem for another day.

Also a problem for another day: changing the conditions under which SMP hands get opened (which can be a point or two lighter than the rule of 20).

Signs this is a good change: this PR adds new functions while shrinking the codebase!


* reorder topics in web interface

* create wrappers for when only North or South must be an unpassed hand

* add comments

* use the new wraps in SMP 2D openers, and fix 2 bugs!

* use new wrappers in forcing 1N topic

* WAI is only for webserver stuff

* print where the .tex file is

* use new wrappers in even more places