More on Lampe responses (#72)
I had stopped work on this because I wasn't sure what to do when responder had a 6-card minor and a 4-card major. Surprise: the answer turns out to be canape bids! So, here's some more work on it. I'm still missing situations for a lot of the follow-ups when opener is a balanced minimum, but it's coming along.

A problem: the version Jordan taught Lee and the version Jordan taught Monique are different! This is the version I got from Lee, which is more artificial but I think clearer and easier to use if you can remember everything.

This name `cannotPreempt2H` is misleading: it's for when your preemptive bids start at 2H, whereas 2D has some conventional meaning (Precision, Benjamin Twos, Mini-Roman, Mexican 2D, etc.). What's the right name for that? I haven't yet found one I'm happy with.


* more work on Lampe: Shape-Showers after opener is an unbalanced minimum but responder is GF

* define the shape-shower follow-ups for 5-3 hands, not sure when to use them yet

* add canape bids

* make `>-` into a function, not an operator

* reorder arguments to addCall to make it more function-like

* pr feedback: doesn't compile right now because some values are unused

* define builds to distinguish 5-3 in the minors further