Make a topic for 4441 hands after an SMP 1C opening (#83)
It's a smaller, more limited topic than I expected, but I think it's separate enough to not be part of a 4C/4D/RKC topic yet. There is a single auction in a single situation that should be modified if you use the alternative responses (1C-2H-3S should become 1C-1N-3C), which felt big enough to create an Action for it (`b1C1Nalt3C`), but not big enough to create separate `Situations` for the two, let alone separate `Topic`s. If someone really cares, they can edit the source code themselves as relevant. 

I think this is the first time I've had nested `<~`s, where the choices available in a later `<~` depend on the choice made in an earlier one. It looks ugly; is there a prettier way to phrase it? Maybe not...


* add the relay bids for 4441 hands with 1C openers

* sort the exported bids from SMP/OneClub.hs

* taking a first stab at a very awkward situation

* get it to compile

* situations for bidding the singletons

* tweak to topic description

* reorder topics in web interface

* more tweaks

* add bids for 4441 hands with a singleton in partner's primary suit

* make a situations to practice when we have 4441 in partner's suit