Add a topic for Cappelletti over weak notrump (#85)
One can argue that I should also put in situations when partner overcalls a major but we have our own suit and bid that instead of asking partner what their minor is. but come on, how often do people actually bid that, rather than mistakenly bidding 3C as "pass or correct" instead of bidding 2N like they should have? It's arguably so rare that not practicing it is good enough.


* start of Cappelletti topic

* fix typos

* rephrase explanation in Forcing 1N

* remember the relay bid after a single-suited hand

* give the alerts a fixed width in the web interface

* add situations to ask what a single-suited hand is

* add Situations for overcalling 2D

* situation for overcalling 2M

* add situations for showing both minors

* add situations asking for partner to bid their second suit

* make the bidding have a fixed width in the web interface, second attempt

* whitespace tweak