Create haskell.yml (#53)
Hopefully this makes sure we can build on both Ubuntu and Mac. The first time took 8ish minutes to build, and the second time took 2 minutes. I hope the 8-minute one was just a one-time thing, and not once per PR. We'll see...


* Create haskell.yml

Hopefully this makes sure we can build on both Ubuntu and Mac...

* crib things from https://github.com/snoyberg/mono-traversable/blob/master/.github/workflows/tests.yml

* more copying of that other repo

* fix bug from refactoring

* add comment to demonstrate that caching works