Implement a suitBid function and use it in the topics (#86)
There is something off about `Action` being an _alias_ for `State Auction ()`, but I can't quite articulate how to fix it. So, I put in a giant warning, and hopefully I'll figure it out later. 

In the meantime, everything still compiles and the code feels simpler even though this PR increases the line count. :shrug: 

I also tried a different caching action in the Github runner. We'll see if it works better; if not, I'll go back to the old one and try to cache stuff better when I merge a PR.


* create suitBid as a helper function

* return the suit, not a string describing the suit

* this works, but isn't great. Maybe Action shouldn't be an alias

* add explanatory comment

* add other comments, too

* remove unnecessary SuitBid instance declaration

* convert the rest of the Forcing 1N topic to use T.suitBid

* use suitBid in the major suit raises topic

* use suitBid in the texas transfers topic

* use suitBid in the Jacoby 2NT topic

* use suitBid in the Jacoby transfers topic

* use suitBid in the Stayman topic

* attempt to make the github actions cache better

* pr feedback