make the suit symbols all line up with each other (#89)
This is super ugly! I wonder if there's a better approach, but I don't see it immediately. Doing this with actual DOM manipulation instead of using `innerHTML` feels both more proper and _way_ more of a pain.

All this effort just to prevent the suits from having ragged alignment. This is so much easier in LaTeX!


* make the suit symbols all line up with each other

* hand heights can be a little smaller

* another attempt