Add bids for takeout doubles (#135)
This works on #125 but does not close it.

I've left out jump advances showing extra strength. Should those be only when you jump to the 2 level, or when the cheapest bid is at the 2 level and you're jumping to the 3 level? Should it be for passed hands, or unpassed hands, or both? I don't want to bother figuring it out right now, so I left a todo.


* start adding bids for takeout doubles, about halfway through

* done defining non-jump advances over takeout doubles

* add TODO for another time

* add todo