Refactor the compile-time assertions (#169)
The original plan here was to assert that every Situation within a Topic had a unique name. Unfortunately, this is not true! To get situation A twice as common as situation B, we use the idiom `wrap [sitA, sitA, sitB]` in multiple places.

but in the meantime, I did find a couple copied-and-pasted situations whose name should have been updated but wasn't! 


* start of compile-time assertions, should build out more

* export a generalized duplicatesOf

* delete the old, complicated way to check for duplicate IDs

* tweak

* get the new assertion up and going, needs cleanup

* no, this doesn't work when wrapping the same situation twice. Rethink

* fix duplicate names in copied-and-pasted situations, whoops

* whitespace tweak

* we no longer use Template Haskell in SupportedTopics.hs directly

* add a compile-time assertion that topic debug strings are unique, too

* also assert that topics have unique debug names

* fix duplicate copied-and-pasted debug string for topic

* whitespace tweak